+++ rust_fsm/examples/circuit_breaker_dsl.rs
@@ -13,8 +13,10 @@ state_machine! {
 
     Closed(Unsuccessful) => Open [SetupTimer],
     Open(TimerTriggered) => HalfOpen,
-    HalfOpen(Successful) => Closed,
-    HalfOpen(Unsuccessful) => Open [SetupTimer],
+    HalfOpen => {
+        Successful => Closed,
+        Unsuccessful => Open [SetupTimer]
+    }
 }
 
 fn main() {