+++ rust_fsm/src/lib.rs
@@ -56,8 +56,10 @@
 //!
 //!     Closed(Unsuccessful) => Open [SetupTimer],
 //!     Open(TimerTriggered) => HalfOpen,
-//!     HalfOpen(Successful) => Closed,
-//!     HalfOpen(Unsuccessful) => Open [SetupTimer],
+//!     HalfOpen => {
+//!         Successful => Closed,
+//!         Unsuccessful => Open [SetupTimer]
+//!     }
 //! }
 //! ```
 //!